+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 6.0
 - HTML Sort Order, Length & Display Column: use "number" input type, program wide
 - Add TipMessage to Food Service Icon in FS_Icons.inc.php
 - HTML standardize Name + ID & Balance headers in Accounts.php, ServeMenus.php, Statements.php & Transactions.php
+- JS update jQuery to v2.2.4 in jquery.js
 
 Changes in 5.9
 --------------